Break core:contract dependency on core:base by moving out Command processing references
